pseudopotential_tuning/all_electron_data/requirements.txt
aiida-core==2.5.1
aio-pika==6.8.1
aiormq==3.3.1
alembic==1.13.1
annotated-types==0.6.0
archive-path==0.4.2
asttokens==2.4.1
async-generator==1.10
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
circus==0.18.0
click==8.1.7
click-spinner==0.1.10
cryptography==42.0.5
decorator==5.1.1
deprecation==2.1.0
disk_objectstore==1.1.0
docstring_parser==0.16
exceptiongroup==1.2.0
executing==2.0.1
graphviz==0.20.3
greenlet==3.0.3
idna==3.7
importlib-metadata==6.11.0
ipython==8.23.0
jedi==0.18.2
Jinja2==3.1.3
kiwipy==0.7.8
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pamqp==2.3.0
paramiko==2.12.0
parso==0.8.4
pexpect==4.9.0
pgsu==0.2.4
plumpy==0.21.10
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyNaCl==1.5.0
pytray==0.3.4
pytz==2021.3
PyYAML==6.0.1
pyzmq==26.0.0
requests==2.31.0
shortuuid==1.0.13
six==1.16.0
SQLAlchemy==2.0.29
stack-data==0.6.3
tabulate==0.8.10
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
upf-to-json==0.9.5
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
